requirements.txt
Django==5.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
requests==2.31.0
requests-oauthlib==1.3.1
